Init Commit
A Re-organization of the GANOS Engine. Currently the SavingIO of Regions has been rebuilt to use Huffman Trees. Ongoing work.

The Huffman tree loses on a few key points. The positioning of Entities which MUST have a unique UUID ( Which gets sterilized so they can keep track of each other ) so their inclusion is wasted since they have to use bytes to represent their own code, as well as their original copy. Consiterations of having entities added on at the end with a positioning offset.