Update #5 Region changes, World Registration, Steps towards Operation
In order of additions

ViewBase - Takes a EntityBase as its permenant target, and collects the regions it should be able to veiw through it, using effects to add/modify entities, blocks and regions seen by it. Every entity with non omipitant AI should have this. Not complete yet but framework is being laid out for it

World - Since regions only exist in one dimension now since having overlapping regions is beyond tedious. Chunks are no more ( Unless we need to optimise Region/Entity/Radius location ), but I feel for now lets avoid unessisary complication. Spawn Block/Entity/Region and Despawn are implemented partially but will need some desigining and work to add in replication support from the Server

Region: Now has an ID as a parameter, ( Of if not will soon ), this will be how by default regions will connect to each other and be stored and located. This will allow for a single Integer ( 4 bytes ) to represent a connection betwen regions

EntityBase: Read Data will now need to be a constructor not just a method due to the finality that UUID should have.

Location: Has new methods like getDirectionX which gives the location X blocks in a given direct, and getDistance, getLocationDifference, ect.