Update #7
Rewrote the engine on the new System
- Entities now have access to the WorldGameState they originated from,
this is to properly facilitate any extra logic added to map changing
methods, such as replication. They also now have access to a
RegionUUID, which allows for reference to their parent region without
specifically having access to it
- Blocks do the same
- World now is nothing more than a data holding object, as it should
- World now handles RegionConnections and Region Spawn locations, which
will be part of the World config.
- Location now uses UUID world instead of World