Update 20 Walkie Talkie
Decided to start adding Functional UI elements, the Walkie Talkie.

As of current it contains support for Channels, Icons, Battery, Messages and Inhabitant, as well as proper transitions.

Current on the To DO list is adding proper Rendering, Message and Speaker support as well as Inhabitant icons, as well as even future TODO Battery stuff