Update #29 Updating Entity Location "Fixing"
Entities when stored in IO have their location expressed as 3 bytes, representing their offset in the current region they are in. This is to save spaces in RegionFiles, but since the same IO is shared for Packets and PlayerFiles, it remains. This location is fixed before they are ready to be spawned. This has been properly implemented.