Update #37 Nearing Designer Block finish Line
Designer blocks need an associated Mesh, AssetLoader needs a way to associate InstanceGridTextures, and the render code needs to be cleaned and ridded of the nulls I added just to make it work.

Future work towards making a world.renderScene() needs to be made, and uniform, since the Client and Server should use the same rendering stuff for blocks and entities.

Transparency still needs to be added :_: