Update #43 FreeTTS support
So on a wim, and a general idea, I implemented FreeTTS support, along with associated Mbrola support to the engine. This allows for up to 5 ( 6 if you count limited ) voices to be used as general effects within Ganos. Who knows, maybe future features could be more audiably associated, or can be used as a part of character dialog