Update #46 Navigation Time!!!
Now, I am adding a utility/blocks/entities that are able to navigate, using some sort of algorithum for path movement.

 The plans are to implement 2, maybe 3 algorithums. Flood, Greedy, and A*, with A* probably being the likely used one, however due to the existance of portals which can make directed movement... unstable, other two algorithums mgiht be useful to add.

Future updates need to be concerned that NavigatableBlocks might be an issue, as they currently have no way to properly update their internal registry to nearby blocks, however a full list of all navigatableblocks is being kept on World, so updating it should be relatively simple. Maybe a "Finalize" world Function