Update #47 Navigation Debug Block
So good news, I deciced that we probably doooooon't want to save our pathfinding network, so for now that stuff is manually generated *Finger gun*

So now with our DebugNavigationBlock being a DesignerBlock, we can change the top plate texture to match its connections, offering 16 posibities! So it will help us visualize the awesomeness!