Update #55 Radient Vs Skylight and Static
Static lights have been implemented, and now there are 2 types of light producers, Radient ( In a direction/s ) or Skylight ( Down and with a Width and Length ).

There appears to be a lighting glitch occurering at certain places, so some "Structure" needs to be made before we can 100% say lighting is completed. But it does look like it functions in a sufficent capacity