Update #56 Stucture and Lighting bug
Good news and slighly bad news!

Good news, we have officially lighting and a basic stucture for lighting tests
Slightly bad news, our light isn't working on the floor when y==1. Either because the spreading algorithum is messing up on BELOW/ABOVE, or because the type of block DebugNavigationBlock isn't working properly