Update #59 Lighting Woes
So lighting works... but getting it properly replicated is being a bit of an issue. As of currenty... lighting isn't properly replicating, and it looks like recalculating light per player may be more expensive than we can currently manage. So for now, despite the fact we've worked so hard on a secular view... if we wanna maintain decent lighting update times we need to relent and force all dynamic/static light to be visible to all players.